server/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
paho-mqtt==1.6.1
python-multipart==0.0.6
